fix(build-output): show stack during CSR bailout warning (#62594)
We should treat the warning for CSR bailout the same as the hard-error,
and show the stack trace for either case.

It might be useful to track down the source of this warning.

We have the stack info already, but we only showed it during the error
logging. We should show it for warnings too.

This is similar to #61200

Closes NEXT-2624