Fixed improper nesting of unordered list inside ordered list on Authentication docs page
On the Building Your Application: Authentication page, the formatting is broken in the "Setting Up Middleware" section because the unordered list is nested inside the ordered list. This is causing the unordered list components to be numbered and throw off the ordering of the top level list.  

I made a tiny change to the markdown to make the top level steps h4 titles, and un-nested the unordered list's