fix false positive test
Previously, this test was passing even for dev, since `(global as any).isDev` referred to a non-existent value (isNextDev is the correct name), never executing the dev part of the test.
Multiple `expect`s were still wrong, as `getRedboxSource` would have never returned the error message.
Because of the lacking type safety on `browser`, the dev test was still not running properly as `elementsByCss` always returns an array