fix another false positive test
Previously, this test was passing even for dev, since `(global as any).isDev` referred to a non-existent value (isNextDev is the correct name), never executing the dev part of the test.

Correctly running the test alters the behavior of later tests without first restoring a patched file to the original content.

This also shows another weakness in our testing, where we sometimes rely on previous test's side effects to run correctly