Add distinction between Development and Production runtime_type for Turbopack (vercel/turbo#7711)
### Description

Currently the development Turbopack runtime is always used as production
builds are not fully implemented yet. This is the first step to having a
production-specific runtime by adding a distinction between Development
and Production for the runtime. This PR does not implement the
production runtime, it only adds the plumbing for adding the production
specific runtime. The runtime implementation will follow in a later PR.