docs: add Lucia for authentication (#63339)
I used Lucia in my last App Router project and it was a super smooth
experience. I would have put it on top of the list, because it is open
source, you keep your user data, it is not a paid service, and has no
lock-in, however, I though I would have to put it below Clerk, because
Clerk is official Vercel partner (?). Anyway, would love to see Lucia in
the list!

Co-authored-by: Sam Ko <sam@vercel.com>