Update a third-parties docs, GoogleTagManager example (#63448)
### What?
If I follow the example in docs then use GTM, it doesn't work when I
send an event in page when mounted using useEffect. It logs a warning:
`@next/third-parties: GTM dataLayer dataLayer does not exist`

### Why?
`window.dataLayer` does not exist when children is rendered, because
children runs faster than GoogleTagManager component. It works in dev
mode, but doesn't work when built, so it can be easily missed.

### How?
It works fine when the order is changed. As you see in example,
GoogleTagManager should come first, higher than children.